ToolSource/pyproject.toml
[project]
name = "aioptim"
version = "1.0.0"
description = "AI backend optimisation tool"
authors = [
    {name = "Lavish Kamal Kumar",email = "lavish.kamal_kumar@kcl.ac.uk"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer (>=0.15.2,<0.16.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "schedule (>=1.2.2,<2.0.0)",
    "prefect (>=3.2.14,<4.0.0)",
    "transformers (>=4.50.1,<5.0.0)",
    "pygithub (>=2.6.1,<3.0.0)",
    "rapidfuzz (>=3.12.2,<4.0.0)",
    "tree-sitter (>=0.24.0,<0.25.0)",
    "tree-sitter-python (>=0.23.6,<0.24.0)",
    "tree-sitter-java (>=0.23.5,<0.24.0)",
    "torch (>=2.6.0,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "colorist (>=1.8.3,<2.0.0)",
    "pytest-cov (>=6.1.1,<7.0.0)",
]

[tool.poetry]
packages = [{include = "aioptim", from = "src"}]


[tool.poetry.scripts]
aioptim = "aioptim.cli.main:app"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = [
    "*/*remote_module_non_scriptable.py",
    "*/tests/*"
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
    # note the use of single quote below to denote "raw" strings in TOML
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
]
